requirements.txt
pyqt
jsonpickle>=2.2.0
mediapipe>=0.8.10
opencv-python>=4.6.0.66
PyAutoGUI>=0.9.53
numpy>=1.23.0